Plane change accuracy fix
Plane changes should include a retrograde component to avoid changing
the total speed, small for a small change and bigger for a large change.
At 90deg, you need to completely zero out your current velocity while
establishing the same velocity in the new direction.

The MechJeb code that we borrowed accounted for this already! but we
were discarding the distinctions between the different components and
putting the whole magnitude of the vector into the normal component of
the burn, because I didn't understand the subtleties of this at the
time. This meant that our total velocity ended up a bit too high, so we
tended to overshoot Minmus, for example.

Now, we preserve the different components and use them for the burn,
which means the normal component will be smaller, with the difference
moved into a new retrograde component. This keeps our total speed the
same, and therefore the apoapsis also the same, so Minmus overshoots
are a thing of the past.